Move HeadlessPluginContext and Canvas3DRenderer to the main Mol* code
